@@ -1,1 +1,1 @@
-// Reexport your entry components here
+// place files you want to import through the `$lib` alias in this folder.
