@@ -1,16 +1,15 @@
-import { vitePreprocess } from '@sveltejs/kit/vite';
-import customAdapter from './custom-adapter/index.js';
+import adapter from '@sveltejs/adapter-auto';
+import {vitePreprocess} from '@sveltejs/kit/vite'
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
-	// for more information about preprocessors
-	preprocess: vitePreprocess(),
-
 	kit: {
-		adapter: customAdapter(), // use the custom adapter
-		// ... other options you may have
-	}
+		// adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
+		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
+		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
+		adapter: adapter()
+	},
+	preprocess: vitePreprocess()
 };
 
 export default config;
